refactor: extract application bootstrap sequence into dedicated module #550
